fix: Relocate ConfigFieldAttribute and related classes to the Runtime directory, and break them into separate files.
