+++ com.playeveryware.eos/Runtime/Core/Config.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a8d289f885733645b739f14c5b34a07
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
