+++ com.playeveryware.eos/Runtime/Core/Config/EOSConfig.cs
