+++ com.playeveryware.eos/Runtime/Core/Config/LogLevelConfig.cs
