+++ com.playeveryware.eos/Runtime/Core/Config/Meta.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc78fbbfa938ad646b5334c2c1a559a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
