+++ com.playeveryware.eos/Runtime/Core/Config/Meta/ButtonField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 765b2de7e1d73ab49ae48f913481dc47
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
