+++ com.playeveryware.eos/Runtime/Core/Config/Meta/ConfigFieldType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee990cbd97ed71c458b0dab903912952
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
