+++ com.playeveryware.eos/Runtime/Core/Config/Meta/DirectoryPathField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5966f0a799ef4c34c9ab39a0879e4e84
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
