+++ com.playeveryware.eos/Runtime/Core/Config/Meta/FilePathField.cs
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2024 PlayEveryWare
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace PlayEveryWare.EpicOnlineServices
+{
+    using System;
+
+    /// <summary>
+    /// This attribute is used to decorate a field member within a config class
+    /// that represents a path to a file.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Field)]
+    public class FilePathField : ConfigFieldAttribute
+    {
+        /// <summary>
+        /// When launching the file selector in the system, use this to
+        /// filter which files can be selected.
+        /// </summary>
+        public string Extension { get; }
+
+        public FilePathField(string label, string extension, string tooltip = null, int group = -1) : base(label,
+            ConfigFieldType.FilePath, tooltip, group)
+        {
+            Extension = extension;
+        }
+    }
+}
