+++ com.playeveryware.eos/Runtime/Core/Config/Meta/FilePathField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1955b7d697ee91c46a1c0f1c2706cbbd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
