+++ com.playeveryware.eos/Runtime/Core/Config/PlatformConfig.cs
